Refactor signup and login validation functions
