+++ src/routes/users-routes.js
@@ -1,81 +1,13 @@
 var express = require('express');
 var passport = require('passport');
+var validateLogin = require('../utils/form-validation').validateLogin;
+var validateUserSignup = require('../utils/form-validation').validateUserSignup;
 var router = express.Router();
-
-var validateUserLogin = function(payload) {
-  var errors = {};
-  var isFormValid = true;
-
-  var username = payload.username;
-  var password = payload.password;
-
-  if (
-    !payload ||
-    typeof username !== 'string' ||
-    username.trim().length === 0
-  ) {
-    isFormValid = false;
-    errors.username = 'Please enter a username';
-  }
-
-  if (
-    !payload ||
-    typeof password !== 'string' ||
-    password.trim().length === 0
-  ) {
-    isFormValid = false;
-    errors.password = 'Please enter a password';
-  }
-
-  return {
-    success: isFormValid,
-    message: !isFormValid ? 'Check the form for errors' : '',
-    errors: errors
-  };
-};
-
-var validateUserSignup = function(payload) {
-  var errors = {};
-  var isFormValid = true;
-
-  var username = payload.username;
-  var password = payload.password;
-  var zipcode = payload.zipcode;
-
-  if (
-    !payload ||
-    typeof username !== 'string' ||
-    username.trim().length === 0
-  ) {
-    isFormValid = false;
-    errors.username = 'Please enter a username';
-  }
-
-  if (
-    !payload ||
-    typeof password !== 'string' ||
-    password.trim().length === 0
-  ) {
-    isFormValid = false;
-    errors.password = 'Please enter a password';
-  }
-
-  if (!payload || typeof zipcode !== 'string' || zipcode.trim().length === 0) {
-    isFormValid = false;
-    errors.zipcode = 'Please enter a zipcode';
-  }
-
-  return {
-    success: isFormValid,
-    message: !isFormValid ? 'Check the form for errors' : '',
-    errors: errors
-  };
-};
 
 // /api/users/login
 // Login for users
 router.post('/login', function(req, res, next) {
-  var validation = validateUserLogin(req.body);
+  var validation = validateLogin(req.body);
   if (!validation.success) {
     return res.status(400).json({
       success: false,
